MNT: Adjust imports after moving eval_results to interface.base
