+++ datalad/core/distributed/clone.py
@@ -25,13 +25,13 @@ from datalad.distribution.dataset import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results
 )
 from datalad.interface.common_opts import (
     location_description,
     reckless_opt,
 )
 from datalad.interface.results import get_status_dict
-from datalad.interface.utils import eval_results
 from datalad.support.annexrepo import AnnexRepo
 from datalad.support.constraints import (
     EnsureKeyChoice,
