+++ datalad/core/local/diff.py
@@ -22,8 +22,8 @@ from datalad.utils import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
-from datalad.interface.utils import eval_results
 
 from datalad.distribution.dataset import (
     Dataset,
