+++ datalad/distributed/create_sibling_github.py
@@ -26,8 +26,8 @@ from datalad.distribution.dataset import datasetmethod
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
-from datalad.interface.utils import eval_results
 from datalad.support.constraints import (
     EnsureChoice,
     EnsureNone,
