+++ datalad/distributed/create_sibling_ria.py
@@ -36,13 +36,13 @@ from datalad.distribution.utils import _yield_ds_w_matching_siblings
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.common_opts import (
     recursion_flag,
     recursion_limit,
 )
 from datalad.interface.results import get_status_dict
-from datalad.interface.utils import eval_results
 from datalad.log import log_progress
 from datalad.support.annexrepo import AnnexRepo
 from datalad.support.constraints import (
