+++ datalad/distributed/export_archive_ora.py
@@ -25,11 +25,11 @@ from datalad.utils import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.results import (
     get_status_dict,
 )
-from datalad.interface.utils import eval_results
 from datalad.support.param import Parameter
 from datalad.support.constraints import (
     EnsureChoice,
