+++ datalad/distributed/export_to_figshare.py
@@ -190,7 +190,7 @@ class ExportToFigshare(Interface):
 
     from datalad.support.param import Parameter
     from datalad.distribution.dataset import datasetmethod
-    from datalad.interface.utils import eval_results
+    from datalad.interface.base import eval_results
     from datalad.distribution.dataset import EnsureDataset
     from datalad.support.constraints import (
         EnsureChoice,
