+++ datalad/distribution/create_sibling.py
@@ -48,6 +48,7 @@ from datalad.distribution.siblings import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.common_opts import (
     annex_group_opt,
@@ -60,7 +61,6 @@ from datalad.interface.common_opts import (
     recursion_flag,
     recursion_limit,
 )
-from datalad.interface.utils import eval_results
 from datalad.support.annexrepo import AnnexRepo
 from datalad.support.constraints import (
     EnsureBool,
