+++ datalad/distribution/update.py
@@ -18,8 +18,10 @@ from os.path import lexists
 import itertools
 
 from datalad.interface.base import Interface
-from datalad.interface.utils import eval_results
-from datalad.interface.base import build_doc
+from datalad.interface.base import (
+    build_doc,
+    eval_results,
+)
 from datalad.interface.results import (
     get_status_dict,
     YieldDatasets
