+++ datalad/interface/shell_completion.py
@@ -14,8 +14,10 @@ __docformat__ = 'restructuredtext'
 
 from .base import Interface
 
-from datalad.interface.base import build_doc
-from datalad.interface.utils import eval_results
+from datalad.interface.base import (
+    build_doc,
+    eval_results,
+)
 from datalad.interface.results import get_status_dict
 from datalad.support.exceptions import CapturedException
