+++ datalad/local/add_archive_content.py
@@ -37,10 +37,10 @@ from datalad.distribution.dataset import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.common_opts import allow_dirty
 from datalad.interface.results import get_status_dict
-from datalad.interface.utils import eval_results
 from datalad.log import (
     log_progress,
     logging,
