+++ datalad/local/check_dates.py
@@ -66,7 +66,7 @@ class CheckDates(Interface):
     timestamps within files of the "git-annex" branch, and (3) the timestamps
     of annotated tags.
     """
-    from datalad.interface.utils import eval_results
+    from datalad.interface.base import eval_results
     import datalad.support.ansi_colors as ac
     from datalad.support.constraints import EnsureChoice, EnsureNone, EnsureStr
     from datalad.support.param import Parameter
