+++ datalad/local/copy_file.py
@@ -17,8 +17,10 @@ from shutil import copyfile
 import sys
 
 from datalad.interface.base import Interface
-from datalad.interface.utils import eval_results
-from datalad.interface.base import build_doc
+from datalad.interface.base import (
+    build_doc,
+    eval_results,
+)
 from datalad.support.constraints import (
     EnsureStr,
     EnsureNone,
