+++ datalad/local/download_url.py
@@ -27,13 +27,13 @@ from datalad.downloaders.providers import Provider
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.common_opts import (
     nosave_opt,
     save_message_opt,
 )
 from datalad.interface.results import get_status_dict
-from datalad.interface.utils import eval_results
 from datalad.support.annexrepo import AnnexRepo
 from datalad.support.constraints import (
     EnsureNone,
