+++ datalad/local/foreach_dataset.py
@@ -33,6 +33,7 @@ from datalad.distribution.dataset import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.common_opts import (
     contains,
@@ -42,7 +43,6 @@ from datalad.interface.common_opts import (
     recursion_limit,
 )
 from datalad.interface.results import get_status_dict
-from datalad.interface.utils import eval_results
 from datalad.support.constraints import (
     EnsureBool,
     EnsureChoice,
