+++ datalad/local/rerun.py
@@ -35,9 +35,9 @@ from datalad.distribution.dataset import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.results import get_status_dict
-from datalad.interface.utils import eval_results
 from datalad.interface.common_opts import jobs_opt
 
 from datalad.support.constraints import (
