+++ datalad/local/run_procedure.py
@@ -31,9 +31,9 @@ from datalad.distribution.dataset import (
 from datalad.interface.base import (
     Interface,
     build_doc,
+    eval_results,
 )
 from datalad.interface.results import get_status_dict
-from datalad.interface.utils import eval_results
 from datalad.support.constraints import EnsureNone
 from datalad.support.exceptions import (
     InsufficientArgumentsError,
