+++ datalad/local/subdatasets.py
@@ -16,10 +16,12 @@ import re
 import os
 import warnings
 
-from datalad.interface.base import Interface
 from datalad.interface.utils import generic_result_renderer
-from datalad.interface.utils import eval_results
-from datalad.interface.base import build_doc
+from datalad.interface.base import (
+    Interface,
+    build_doc,
+    eval_results,
+)
 from datalad.interface.results import get_status_dict
 from datalad.support.constraints import (
     EnsureStr,
