+++ datalad/local/wtf.py
@@ -19,8 +19,10 @@ from functools import partial
 from collections import OrderedDict
 
 
-from datalad.interface.base import Interface
-from datalad.interface.base import build_doc
+from datalad.interface.base import (
+    Interface,
+    build_doc,
+)
 from datalad.utils import (
     ensure_unicode,
     getpwd,
@@ -363,7 +365,7 @@ class WTF(Interface):
 
     from datalad.support.param import Parameter
     from datalad.distribution.dataset import datasetmethod
-    from datalad.interface.utils import eval_results
+    from datalad.interface.base import eval_results
     from datalad.distribution.dataset import EnsureDataset
     from datalad.support.constraints import EnsureNone, EnsureChoice
